requirements.txt
GitPython>=1.0.2
colorama>=0.3.7
termcolor>=1.1.0
docopt>=0.6.2
six>=1.10.0
